The Scene And Player Readin Logic

+++ Src/Assets/Common/Data/Space/PlayerSpacing.cs
@@ -8,6 +8,6 @@ using System.Threading.Tasks;
 public class PlayerSpacing
 {
     //每个玩家的二维坐标
-    public List<int> Player_X;
-    public List<int> Player_Y;
+    public List<float> Player_X;
+    public List<float> Player_Y;
 }
